+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/configuration/SimpleTemplateCompilerConfig.java
@@ -28,6 +28,7 @@ public class SimpleTemplateCompilerConfig extends TemplateCompilerConfig {
         return "SimpleTemplateCompilerConfig{" +
                 "@type='" + type_ + '\'' +
                 ", name='" + name + '\'' +
+                ", fullQualifiedName='" + fullQualifiedName + '\'' +
                 ", template='" + template + '\'' +
                 ", package_='" + package_ + '\'' +
                 ", bindings='" + bindings + '\'' +