+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/configuration/TemplateCompilerConfig.java
@@ -14,6 +14,8 @@ abstract public class TemplateCompilerConfig {
     public String type_;
     @JsonProperty("name")
     public String name;
+    @JsonProperty("fullQualifiedName")
+    public String fullQualifiedName;
     @JsonProperty("package")
     public String package_;