+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/compiler/expansion/CompilerExpansionBuilder.java
@@ -87,7 +87,6 @@ public class CompilerExpansionBuilder {
 
         JavaFile myfile = compilerUtil.specWithComment(theInterface, templateName, packge, stackTraceElement);
 
-        System.out.println("******* " + fileName + " " + packge);
         return new SpecificationFile(myfile, directory, fileName, packge);
     }