+++ modules-template/prov-template-compiler/src/main/java/org/openprovenance/prov/template/emitter/PoetParser.java
@@ -76,7 +76,7 @@ public class PoetParser {
     }
 
     private List<Statement> parse(CodeBlock codeBlock) {
-        System.out.println("******** Poet code block no longer handled for export to python ****");
+      //  System.out.println("******** Poet code block no longer handled for export to python ****");
         return List.of();
     }