+++ modules-template/prov-template/src/main/java/module-info.java
@@ -8,6 +8,7 @@ module prov.template {
     requires prov.model;
     requires java.xml;
 
+    exports org.openprovenance.prov.template.utils;
     exports org.openprovenance.prov.template.json;
     exports org.openprovenance.prov.template.json.deserializer;
     exports org.openprovenance.prov.template.expander;