+++ modules-template/prov-template/src/main/java/org/openprovenance/prov/template/utils/VariableCheckException.java
@@ -1,0 +1,6 @@
+package org.openprovenance.prov.template.utils;
+
+public class VariableCheckException extends RuntimeException {
+    public VariableCheckException(String s) {
+    }
+}